refactor: centralize skill security utilities
